Wrapping up things. Implemented PR suggestions #11.
